+++ processor/src/main/java/router/fu/processor/Generator.java
@@ -485,7 +485,10 @@ final class Generator
         }
         else
         {
-          sb.append( "((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})+)" );
+          final ParameterDescriptor parameter = (ParameterDescriptor) part;
+          sb.append( "((?:[a-zA-Z0-9\\-_\\$\\.\\+!\\*'\\(\\)\\,\\\"~]|%[A-F0-9]{2})" );
+          sb.append( parameter.isOptional() ? "*" : "+" );
+          sb.append( ")" );
         }
       }
       if ( !route.isPartialMatch() )